Combine test-real-values.js into test-versions.js
